refactor: 내부 HTTP self-call을 직접 메서드 호출로 전환
Order→Product, Order→Wishlist, Funding→Wishlist 3개 경로의
HTTP self-call(@HttpExchange)을 제거하고 bc/shared Port +
bc/catalog Adapter를 통한 직접 호출로 전환.
- ProductSnapshotPort, WishlistItemSnapshotPort를 bc/shared로 이동
- HTTP client/adapter/config 7개 파일 삭제 (-216줄)
