--- bc/core/src/main/java/app/giftify/funding/adpater/outbound/FundingApiClientConfig.java
@@ -1,27 +1,0 @@
-package app.giftify.funding.adpater.outbound;
-
-import app.giftify.funding.application.outbound.WishlistItemSnapshotApiClient;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestClient;
-import org.springframework.web.client.support.RestClientAdapter;
-import org.springframework.web.service.invoker.HttpServiceProxyFactory;
-
-@Configuration
-public class FundingApiClientConfig {
-    @Bean
-    public WishlistItemSnapshotApiClient wishlistItemSnapshotApiClient(
-            RestClient.Builder builder,
-            @Value("${app.service.catalog.url:http://localhost:8080}") String baseUrl
-    ) {
-        RestClient restClient = builder
-                .baseUrl(baseUrl)
-                .build();
-
-        return HttpServiceProxyFactory
-                .builderFor(RestClientAdapter.create(restClient))
-                .build()
-                .createClient(WishlistItemSnapshotApiClient.class);
-    }
-}
