+++ bc/shared/src/main/java/app/giftify/shared/domain/port/ProductSnapshotPort.java
@@ -1,10 +1,10 @@
-package app.giftify.order.application.outbound.port;
+package app.giftify.shared.domain.port;
 
 import app.giftify.shared.domain.vo.ProductSnapshot;
 
 import java.util.List;
 import java.util.Map;
 
-public interface ProductPort {
+public interface ProductSnapshotPort {
     Map<Long, ProductSnapshot> getProductSnapshots(List<Long> productIds);
 }
